pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-cost-telemetry"
version = "0.1.0"
description = "Agent-first, cost-native telemetry SDK for AI apps"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "prometheus-client",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
    "mypy",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
